@vr009 2022-05-24
Tried to speed up `select` based on this [thread](https://groups.google.com/g/golang-dev/c/jX4oQEls3uk). 
For a single Benchmark the difference is obvious, but for integration tests with a tarantool instance it is not so clear.