@oleg-jukovec 2022-05-25
> Tried to speed up `select` based on this [thread](https://groups.google.com/g/golang-dev/c/jX4oQEls3uk). For a simple Benchmark the difference between the typical construction (without default cases) and this one, is obvious, but for integration tests with a tarantool instance it is not so clear.

An overhead of communicating with the Tarantool instance is much bigger than the effect of such optimizations. It's expected results.