@vr009 2022-05-25
> An overhead of communicating with the Tarantool instance is much bigger than the effect of such optimizations. It's expected results.

Yes, that seems to be truth, maybe we should consider the creating some mock package for tarantool for getting more clean results. @Totktonada your thoughts?