@oleg-jukovec 2022-05-28
> I like the first one due to its native "go" spirit. Also I don't see the way to implement the second option without repeating the same code for every type of request object.

Go has examples of using `WithContext` method. I've not said that we should use this approach. But the links can be useful for discussions.

- https://pkg.go.dev/net/http#NewRequestWithContext
- https://pkg.go.dev/net/http#Request.Context
- https://pkg.go.dev/net/http#Request.WithContext
- https://pkg.go.dev/golang.org/x/sync/errgroup#WithContext