@vr009 2022-06-06
> Go has examples of using WithContext method. I've not said that we should use this approach. But the links can be useful for the discussions.

Thank you for the links, I've missed that http package uses the similar approach with requests objects.
In any case I still see the proposed option as an acceptable one.