@oleg-jukovec 2022-06-20
1. We create a context with 1s timeout (or deadline after 1 second).
2. We call `fut := Connection.DoAsync(ctx, req)`.
3. We do something for 10 seconds or minutes...
4. We call fut.Get().

What we should got as a result if a Connection object receive response in `Connection.writer` after 5 seconds?