@oleg-jukovec 2022-06-20
> When we call fut.Get(), we got closed channel `fut.done` (that happens when we receive a response from instance and fill the Response field of a Future, see `SetResponse()`) and filled `context.Done`. At first we check if the response is come (if fut.done is closed or got a value) and return from the `wait()` call.

Yes, in the current implementation. But shouldn't the request be canceled after 1s?