@vr009 2022-06-20
> Yes, in the current implementation. But shouldn't the request be canceled after 1s?

Seems that we should, I can fix it.