@oleg-jukovec 2022-06-25
As an alternative way we can do something very-very simple and obvious to handle outdated requests with a context instead of `contexts()` goroutine. Something like this:

```Go
--- a/connection.go
+++ b/connection.go
@@ -334,7 +334,7 @@ func Connect(addr string, opts Opts) (conn *Connection, err error) {
     if conn.opts.Timeout > 0 {
         go conn.timeouts()
     }
-    go conn.contexts()
+    //go conn.contexts()
 
     // TODO: reload schema after reconnect.
     if !conn.opts.SkipSchema {
@@ -858,6 +858,15 @@ func (conn *Connection) send(ctx context.Context, req Request) *Future {
         return fut
     }
     conn.putFuture(fut, req)
+    if req.Ctx() != nil {
+        go func() {
+            select {
+            case <-req.Ctx().Done():
+                conn.futCanceler.Cancel(fut, ContextDoneErr)
+            case <-fut.WaitChan():
+            }
+        }()
+    }
     return fut
 }
```
and do the optimization later with well-designed benchmarks.

But this does not change the fact that the current code should be well tested. For example, what happens if conn.futCanceler.Cancel() will be executed in parallel? I think it will be an error here:

https://github.com/tarantool/go-tarantool/blob/7116a33aa26bb7e2390a59a20f555ff3d120387d/future.go#L172-L173